refactor: improve the generation of tracked constants, move them to a separate file, rename them, and name the point as `TRACK_FILES`
