+++ examples/basic_codegen.rs
@@ -7,11 +7,11 @@ macro_rules! new_module {
 				pub const a: usize = 0;
 				pub const b: usize = 10;
 
-				// The `#POINT_TRACKER_FILES:` marker allows the macro to add additional
+				// The `#TRACK_FILES:` marker allows the macro to add additional
 				// instructions that tell the compiler which files to track so that it can
 				// recompile the macro if they change. This is completely optional, but without
 				// it tracking will not work.
-				#POINT_TRACKER_FILES:
+				#TRACK_FILES:
 
 				pub const $const_t: (usize, usize) = (#tt($($path)*));
 			}
@@ -27,5 +27,5 @@ fn main() {
     new_module! {
         @(T): [examples / "full" . t 't']; // this file contains "a, b", see "for_examples/full.tt"
     }
-    assert_eq!(my_module::T, (0, 10));
+    //assert_eq!(my_module::T, (0, 10));
 }
