+++ src/lib.rs
@@ -68,6 +68,7 @@ assert_eq!(buf, "Welcome, Ferris. Your score is 100500!");
 extern crate alloc;
 extern crate proc_macro;
 
+use crate::points::file_dep_tracker::FileDepTracker;
 use crate::trees::null::make_null_group;
 use crate::trees::sq_err;
 use crate::{
@@ -81,9 +82,7 @@ use crate::{
 };
 use core::slice::IterMut;
 use proc_macro::TokenStream;
-use proc_macro2::{Delimiter, Group, Span, TokenStream as TokenStream2, TokenTree as TokenTree2};
-use quote::{format_ident, quote};
-use std::path::Path;
+use proc_macro2::{Group, TokenStream as TokenStream2, TokenTree as TokenTree2};
 
 /// Components, templates, code for the search
 /// and final construction of trees.
@@ -116,119 +115,9 @@ pub(crate) mod exprs {
 /// Code component of macros.
 pub(crate) mod include;
 
-pub(crate) struct PointTrack<'tk> {
-    prefix_token: &'tk mut TokenTree2,
-    name_token: &'tk mut TokenTree2,
-    data_token: &'tk mut TokenTree2,
-    appends_files: usize,
-    globalposnum: usize,
-}
-
-impl<'tk> PointTrack<'tk> {
-    #[inline]
-    pub const fn new(
-        globalposnum: usize,
-        prefix_token: &'tk mut TokenTree2,
-        name_token: &'tk mut TokenTree2,
-        data_token: &'tk mut TokenTree2,
-    ) -> Self {
-        Self {
-            prefix_token,
-            name_token,
-            data_token,
-            appends_files: 0,
-            globalposnum,
-        }
-    }
-
-    #[inline]
-    pub fn prefix_span(&self) -> Span {
-        self.prefix_token.span()
-    }
-
-    #[inline]
-    pub fn name_span(&self) -> Span {
-        self.name_token.span()
-    }
-
-    #[inline]
-    pub fn data_span(&self) -> Span {
-        self.data_token.span()
-    }
-
-    #[inline]
-    pub const fn is_rewritten(&self) -> bool {
-        self.appends_files > 0
-    }
-
-    pub fn into_token_tree2(self) -> Option<(usize, TokenTree2)> {
-        match self.appends_files {
-            0 => None,
-            appends_files => {
-                let data_span = self.data_span();
-
-                Some((
-                    appends_files,
-                    std::mem::replace(self.data_token, make_null_group(data_span)),
-                ))
-            }
-        }
-    }
-
-    pub fn append_track_file(&mut self, path: &Path) {
-        let name_const = format_ident!(
-            "_TRACKER_FILE_NUM_{}",
-            self.globalposnum + self.appends_files
-        );
-
-        let path = format!("../{}", path.display());
-        let ts2 = TokenStream2::from_iter(quote! {
-            /// This is a file tracker point, automatically generated by `#POINT_TRACKER_FILES;`
-            const #name_const: &'static [u8] = include_bytes!(#path) as &[_];
-        });
-
-        self.append_track_files_ts(ts2)
-    }
-
-    pub fn append_track_files_ts(&mut self, ts2: TokenStream2) {
-        let data_span = self.data_span();
-        let is_initappendfiles = self.appends_files == 0;
-        self.appends_files += 1;
-
-        let mut ngroup = Group::new(Delimiter::None, ts2);
-        ngroup.set_span(data_span);
-
-        if is_initappendfiles {
-            *self.data_token = ngroup.into();
-        } else {
-            match &mut self.data_token {
-                TokenTree2::Group(group) => {
-                    let mut new_group: Vec<TokenTree2> = group.stream().into_iter().collect();
-                    new_group.push(ngroup.into());
-
-                    let mut ngroup =
-                        Group::new(Delimiter::None, TokenStream2::from_iter(new_group));
-                    ngroup.set_span(data_span);
-
-                    *self.data_token = ngroup.into();
-                }
-                _ => panic!(
-                    "Undefined behavior reported in `PointTrack`, someone redefined `TokenTree2`, expected `TokenTree2::Group`"
-                ),
-            }
-        }
-    }
-}
-
-impl<'tk> Drop for PointTrack<'tk> {
-    fn drop(&mut self) {
-        if !self.is_rewritten() {
-            let data_span = self.data_span();
-            *self.data_token = make_null_group(data_span);
-        }
-        *self.prefix_token = make_null_group(self.prefix_token.span());
-        *self.name_token = make_null_group(self.prefix_token.span());
-    }
+/// Special code points that override macro behavior
+pub(crate) mod points {
+    pub mod file_dep_tracker;
 }
 
 /// The task of the function is to find a group with the desired macro
@@ -238,7 +127,7 @@ impl<'tk> Drop for PointTrack<'tk> {
 fn autoinject_tt_in_group<'tk, 'gpsn>(
     globalposnum: &'gpsn mut usize,
     mut iter: IterMut<'tk, TokenTree2>,
-    point_track_file: &'_ mut Option<PointTrack<'tk>>,
+    point_track_file: &'_ mut Option<FileDepTracker<'tk>>,
 ) -> SearchGroup {
     'sbegin: while let Some(mut m_punct) = iter.next() {
         'match_m_punct: loop {
@@ -283,12 +172,12 @@ fn autoinject_tt_in_group<'tk, 'gpsn>(
                                     return [ident.span()]: "`:` was expected."
                                 }
                             }
-                            ident if ident == "POINT_TRACKER_FILES" => {
+                            ident if ident == "TRACK_FILES" || ident == "POINT_TRACKER_FILES" => {
                                 if let Some(m_punct2) = iter.next()
                                     && let TokenTree2::Punct(punct2) = m_punct2
                                     && punct2.as_char() == ':'
                                 {
-                                    *point_track_file = Some(PointTrack::new(
+                                    *point_track_file = Some(FileDepTracker::new(
                                         *globalposnum,
                                         m_punct,
                                         m_ident,
@@ -306,7 +195,7 @@ fn autoinject_tt_in_group<'tk, 'gpsn>(
                                 macro_rule_include::<InjectTT>
                                     as fn(
                                         &Group,
-                                        Option<&mut PointTrack<'tk>>,
+                                        Option<&mut FileDepTracker<'tk>>,
                                     )
                                         -> TreeResult<TokenTree2>
                             }
@@ -336,7 +225,7 @@ fn autoinject_tt_in_group<'tk, 'gpsn>(
                             }
 
                             _ => sq_err! {
-                                return [ident.span()]: "Undefined action to include data in macro or change its behavior, expected macro data type: `tt`, `ctt`, `arr`, `str`, or marker: `#AS_IS:`, `#POINT_TRACKER_FILES:`, or stop parsing macro via `#break;`."
+                                return [ident.span()]: "Undefined action to include data in macro or change its behavior, expected macro data type: `tt`, `ctt`, `arr`, `str`, or marker: `#AS_IS:`, `#TRACK_FILES:`, or stop parsing macro via `#break;`."
                             },
                         };
 
@@ -363,14 +252,14 @@ fn autoinject_tt_in_group<'tk, 'gpsn>(
                     let mut prefixgroup;
                     let mut namegroup;
                     let mut datagroup;
-                    #[allow(clippy::manual_map)] // see ngroup
+                    #[allow(clippy::manual_map)]
                     let mut ptf = match point_track_file {
                         Some(point_track_file) => Some({
                             prefixgroup = make_null_group(point_track_file.prefix_span());
                             namegroup = make_null_group(point_track_file.name_span());
                             datagroup = make_null_group(point_track_file.data_span());
 
-                            PointTrack::new(
+                            FileDepTracker::new(
                                 *globalposnum,
                                 &mut prefixgroup,
                                 &mut namegroup,
@@ -423,9 +312,9 @@ fn autoinject_tt_in_group<'tk, 'gpsn>(
 /// * `#arr(...)` or `#array(...)` — Injects content as a `&'static [u8]`.
 ///
 /// ### Control & Optimization:
-/// * `#POINT_TRACKER_FILES:` — Enables incremental compilation tracking for all included files.
+/// * `#TRACK_FILES:` — Enables incremental compilation tracking for all included files.
 /// * `#AS_IS:` or `#break;` — Stops parsing the rest of the macro input to save compile time.
-/// * `-#` — Escapes the `#` symbol so it is ignored by `inject!` (e.g., for attributes).
+/// * `-#` — Escapes the `#` symbol so it is ignored by `inject!`.
 ///
 /// ## Example: Template-like injection
 /// ```rust
@@ -455,7 +344,7 @@ fn autoinject_tt_in_group<'tk, 'gpsn>(
 ///                 pub const b: usize = 10;
 ///
 ///                 // Track changes in external files for incremental compilation
-///                 #POINT_TRACKER_FILES:
+///                 #TRACK_FILES:
 ///
 ///                 pub const $const_t: (usize, usize) = (#tt($($path)*));
 ///             }
